src/ml-models/requirements.txt
keras>=2.8.0
matplotlib>=3.5.1
numpy>=1.22.2
paho_mqtt>=1.6.1
pandas>=1.4.1
scikit_learn>=1.1.3
scipy>=1.8.0
tensorflow>=2.8.0
pdoc>=12.2.0